// src/lib/db-types.ts
/**
 * This file was generated by kysely-codegen.
 * Please do not edit it manually.
 */

import type { ColumnType } from "kysely";

export type Generated<T> = T extends ColumnType<infer S, infer I, infer U>
  ? ColumnType<S, I | undefined, U>
  : ColumnType<T, T | undefined, T>;

export interface Albums {
  author_id: number;
  id: Generated<number>;
  name: string;
  release_date: number;
}

export interface Authors {
  bio: string | null;
  id: Generated<number>;
  name: string;
}

export interface Playlists {
  id: Generated<number>;
  name: string;
  user_id: number;
}

export interface PlaylistsSongs {
  id: Generated<number>;
  playlist_id: number;
  song_id: number;
}

export interface Songs {
  album_id: number;
  duration: number;
  id: Generated<number>;
  name: string;
}

export interface Users {
  email: string;
  id: Generated<number>;
  password: string;
  username: string;
}

export interface DB {
  albums: Albums;
  authors: Authors;
  playlists: Playlists;
  playlists_songs: PlaylistsSongs;
  songs: Songs;
  users: Users;
}